Use amx parameter to prevent flash memory problems
Home Assistant can write new maxCurrent value so often that the flash memory of charger can be broken.
https://github.com/goecharger/go-eCharger-API-v1/blob/master/go-eCharger%20API%20v1%20EN.md
amx 	uint8_t 	Ampere value for the PWM signaling in whole ampere of 6-32A Will not be written on flash but acts like you set amp instead. Only on the next reboot, the amp value will be restored to the last value set with amp. Recommended for PV charging.